// apps/backend/.medusa/types/remote-query-entry-points.d.ts
export type Maybe<T> = T | null;
// ...
export type Scalars = {
  ID: { input: string; output: string; }
  String: { input: string; output: string; }
  Boolean: { input: boolean; output: boolean; }
  Int: { input: number; output: number; }
  Float: { input: number; output: number; }
  DateTime: { input: Date | string; output: Date | string; }
  JSON: { input: Record<string, unknown>; output: Record<string, unknown>; }
};
// ...
export type InventoryItem = {
  __typename?: 'InventoryItem';
  id: Scalars['ID']['output'];
  created_at: Scalars['DateTime']['output'];
  updated_at: Scalars['DateTime']['output'];
  deleted_at?: Maybe<Scalars['DateTime']['output']>;
  sku?: Maybe<Scalars['String']['output']>;
  origin_country?: Maybe<Scalars['String']['output']>;
  hs_code?: Maybe<Scalars['String']['output']>;
  mid_code?: Maybe<Scalars['String']['output']>;
  material?: Maybe<Scalars['String']['output']>;
  weight?: Maybe<Scalars['Int']['output']>;
  length?: Maybe<Scalars['Int']['output']>;
  height?: Maybe<Scalars['Int']['output']>;
  width?: Maybe<Scalars['Int']['output']>;
  requires_shipping: Scalars['Boolean']['output'];
  description?: Maybe<Scalars['String']['output']>;
  title?: Maybe<Scalars['String']['output']>;
  thumbnail?: Maybe<Scalars['String']['output']>;
  metadata?: Maybe<Scalars['JSON']['output']>;
  inventory_levels?: Maybe<Array<Maybe<InventoryLevel>>>;
  variant_link?: Maybe<Array<Maybe<LinkProductVariantInventoryItem>>>;
  variants?: Maybe<Array<Maybe<ProductVariant>>>;
};
// ...
export enum ProductStatus {
  Draft = 'draft',
  Proposed = 'proposed',
  Published = 'published',
  Rejected = 'rejected'
}

export type Product = {
  __typename?: 'Product';
  id: Scalars['ID']['output'];
  title: Scalars['String']['output'];
  handle: Scalars['String']['output'];
  subtitle?: Maybe<Scalars['String']['output']>;
  description?: Maybe<Scalars['String']['output']>;
  is_giftcard: Scalars['Boolean']['output'];
  status: ProductStatus;
  thumbnail?: Maybe<Scalars['String']['output']>;
  width?: Maybe<Scalars['Float']['output']>;
  weight?: Maybe<Scalars['Float']['output']>;
  length?: Maybe<Scalars['Float']['output']>;
  height?: Maybe<Scalars['Float']['output']>;
  origin_country?: Maybe<Scalars['String']['output']>;
  hs_code?: Maybe<Scalars['String']['output']>;
  mid_code?: Maybe<Scalars['String']['output']>;
  material?: Maybe<Scalars['String']['output']>;
  collection?: Maybe<ProductCollection>;
  collection_id?: Maybe<Scalars['String']['output']>;
  categories?: Maybe<Array<Maybe<ProductCategory>>>;
  type?: Maybe<ProductType>;
  type_id?: Maybe<Scalars['String']['output']>;
  tags: Array<ProductTag>;
  variants: Array<ProductVariant>;
  options: Array<ProductOption>;
  images: Array<ProductImage>;
  discountable?: Maybe<Scalars['Boolean']['output']>;
  external_id?: Maybe<Scalars['String']['output']>;
  created_at: Scalars['DateTime']['output'];
  updated_at: Scalars['DateTime']['output'];
  deleted_at?: Maybe<Scalars['DateTime']['output']>;
  metadata?: Maybe<Scalars['JSON']['output']>;
  sales_channels_link?: Maybe<Array<Maybe<LinkProductSalesChannel>>>;
  sales_channels?: Maybe<Array<Maybe<SalesChannel>>>;
  book_link?: Maybe<LinkProductProductBookModuleBook>;
  book?: Maybe<Book>;
};

export type ProductVariant = {
  __typename?: 'ProductVariant';
  id: Scalars['ID']['output'];
  title: Scalars['String']['output'];
  sku?: Maybe<Scalars['String']['output']>;
  barcode?: Maybe<Scalars['String']['output']>;
  ean?: Maybe<Scalars['String']['output']>;
  upc?: Maybe<Scalars['String']['output']>;
  allow_backorder: Scalars['Boolean']['output'];
  manage_inventory: Scalars['Boolean']['output'];
  requires_shipping: Scalars['Boolean']['output'];
  hs_code?: Maybe<Scalars['String']['output']>;
  origin_country?: Maybe<Scalars['String']['output']>;
  mid_code?: Maybe<Scalars['String']['output']>;
  material?: Maybe<Scalars['String']['output']>;
  weight?: Maybe<Scalars['Float']['output']>;
  length?: Maybe<Scalars['Float']['output']>;
  height?: Maybe<Scalars['Float']['output']>;
  width?: Maybe<Scalars['Float']['output']>;
  options: Array<ProductOptionValue>;
  metadata?: Maybe<Scalars['JSON']['output']>;
  product?: Maybe<Product>;
  product_id?: Maybe<Scalars['String']['output']>;
  variant_rank?: Maybe<Scalars['Int']['output']>;
  created_at: Scalars['DateTime']['output'];
  updated_at: Scalars['DateTime']['output'];
  deleted_at?: Maybe<Scalars['DateTime']['output']>;
  inventory_items?: Maybe<Array<Maybe<LinkProductVariantInventoryItem>>>;
  inventory?: Maybe<Array<Maybe<InventoryItem>>>;
  price_set_link?: Maybe<LinkProductVariantPriceSet>;
  price_set?: Maybe<PriceSet>;
  order_items?: Maybe<Array<Maybe<OrderLineItem>>>;
};

export type ProductCategory = {
  __typename?: 'ProductCategory';
  id: Scalars['ID']['output'];
  name: Scalars['String']['output'];
  description: Scalars['String']['output'];
  handle: Scalars['String']['output'];
  is_active: Scalars['Boolean']['output'];
  is_internal: Scalars['Boolean']['output'];
  rank: Scalars['Int']['output'];
  metadata?: Maybe<Scalars['JSON']['output']>;
  parent_category?: Maybe<ProductCategory>;
  parent_category_id?: Maybe<Scalars['String']['output']>;
  category_children: Array<ProductCategory>;
  products: Array<Product>;
  created_at: Scalars['DateTime']['output'];
  updated_at: Scalars['DateTime']['output'];
  deleted_at?: Maybe<Scalars['DateTime']['output']>;
};

export type ProductTag = {
  __typename?: 'ProductTag';
  id: Scalars['ID']['output'];
  value: Scalars['String']['output'];
  metadata?: Maybe<Scalars['JSON']['output']>;
  products?: Maybe<Array<Maybe<Product>>>;
};

export type ProductCollection = {
  __typename?: 'ProductCollection';
  id: Scalars['ID']['output'];
  title: Scalars['String']['output'];
  handle: Scalars['String']['output'];
  metadata?: Maybe<Scalars['JSON']['output']>;
  created_at: Scalars['DateTime']['output'];
  updated_at: Scalars['DateTime']['output'];
  deleted_at?: Maybe<Scalars['DateTime']['output']>;
  products?: Maybe<Array<Maybe<Product>>>;
};

export type ProductType = {
  __typename?: 'ProductType';
  id: Scalars['ID']['output'];
  value: Scalars['String']['output'];
  metadata?: Maybe<Scalars['JSON']['output']>;
  created_at: Scalars['DateTime']['output'];
  updated_at: Scalars['DateTime']['output'];
  deleted_at?: Maybe<Scalars['DateTime']['output']>;
};

export type ProductOption = {
  __typename?: 'ProductOption';
  id: Scalars['ID']['output'];
  title: Scalars['String']['output'];
  product?: Maybe<Product>;
  product_id?: Maybe<Scalars['String']['output']>;
  values: Array<ProductOptionValue>;
  metadata?: Maybe<Scalars['JSON']['output']>;
  created_at: Scalars['DateTime']['output'];
  updated_at: Scalars['DateTime']['output'];
  deleted_at?: Maybe<Scalars['DateTime']['output']>;
};

export type ProductImage = {
  __typename?: 'ProductImage';
  id: Scalars['ID']['output'];
  url: Scalars['String']['output'];
  metadata?: Maybe<Scalars['JSON']['output']>;
  created_at: Scalars['DateTime']['output'];
  updated_at: Scalars['DateTime']['output'];
  deleted_at?: Maybe<Scalars['DateTime']['output']>;
};

export type ProductOptionValue = {
  __typename?: 'ProductOptionValue';
  id: Scalars['ID']['output'];
  value: Scalars['String']['output'];
  option?: Maybe<ProductOption>;
  option_id?: Maybe<Scalars['String']['output']>;
  metadata?: Maybe<Scalars['JSON']['output']>;
  created_at: Scalars['DateTime']['output'];
  updated_at: Scalars['DateTime']['output'];
  deleted_at?: Maybe<Scalars['DateTime']['output']>;
};

export type PriceSet = {
  __typename?: 'PriceSet';
  id: Scalars['ID']['output'];
  prices?: Maybe<Array<Maybe<Price>>>;
  calculated_price?: Maybe<CalculatedPriceSet>;
  variant_link?: Maybe<LinkProductVariantPriceSet>;
  variant?: Maybe<ProductVariant>;
  shipping_option_link?: Maybe<LinkShippingOptionPriceSet>;
  shipping_option?: Maybe<ShippingOption>;
};
// ...
export type SalesChannel = {
  __typename?: 'SalesChannel';
  id: Scalars['ID']['output'];
  name: Scalars['String']['output'];
  description?: Maybe<Scalars['String']['output']>;
  is_disabled: Scalars['Boolean']['output'];
  created_at: Scalars['DateTime']['output'];
  metadata?: Maybe<Scalars['JSON']['output']>;
  updated_at: Scalars['DateTime']['output'];
  deleted_at?: Maybe<Scalars['DateTime']['output']>;
  products_link?: Maybe<Array<Maybe<LinkProductSalesChannel>>>;
  api_keys_link?: Maybe<Array<Maybe<LinkPublishableApiKeySalesChannel>>>;
  orders?: Maybe<Array<Maybe<Order>>>;
  locations_link?: Maybe<Array<Maybe<LinkSalesChannelStockLocation>>>;
  stock_locations?: Maybe<Array<Maybe<StockLocation>>>;
};
// ...
export type Book = {
  __typename?: 'Book';
  id: Scalars['ID']['output'];
  audioVersionUrl?: Maybe<Scalars['String']['output']>;
  dimensions?: Maybe<Scalars['String']['output']>;
  ebookVersionEbook?: Maybe<Scalars['String']['output']>;
  isbn?: Maybe<Scalars['String']['output']>;
  pagesCount?: Maybe<Scalars['Int']['output']>;
  publishDate?: Maybe<Scalars['String']['output']>;
  publisher?: Maybe<Scalars['String']['output']>;
  raw_publishDate?: Maybe<Scalars['JSON']['output']>;
  created_at: Scalars['DateTime']['output'];
  updated_at: Scalars['DateTime']['output'];
  deleted_at?: Maybe<Scalars['DateTime']['output']>;
  author_link?: Maybe<LinkBookModuleBookAuthorModuleAuthor>;
  author?: Maybe<Author>;
  language_link?: Maybe<LinkBookModuleBookLanguageModuleLanguage>;
  language?: Maybe<Language>;
  product_link?: Maybe<LinkProductProductBookModuleBook>;
  product?: Maybe<Product>;
  quote_link?: Maybe<LinkQuoteModuleQuoteBookModuleBook>;
  quote?: Maybe<Quote>;
};
// ...
export type LinkProductSalesChannel = {
  __typename?: 'LinkProductSalesChannel';
  product_id: Scalars['String']['output'];
  sales_channel_id: Scalars['String']['output'];
  product?: Maybe<Product>;
  sales_channel?: Maybe<SalesChannel>;
  createdAt: Scalars['String']['output'];
  updatedAt: Scalars['String']['output'];
  deletedAt?: Maybe<Scalars['String']['output']>;
};

export type LinkProductVariantInventoryItem = {
  __typename?: 'LinkProductVariantInventoryItem';
  variant_id: Scalars['String']['output'];
  inventory_item_id: Scalars['String']['output'];
  required_quantity: Scalars['Int']['output'];
  variant?: Maybe<Product>;
  createdAt: Scalars['String']['output'];
  updatedAt: Scalars['String']['output'];
  deletedAt?: Maybe<Scalars['String']['output']>;
};

export type LinkProductVariantPriceSet = {
  __typename?: 'LinkProductVariantPriceSet';
  variant_id: Scalars['String']['output'];
  price_set_id: Scalars['String']['output'];
  variant?: Maybe<Product>;
  createdAt: Scalars['String']['output'];
  updatedAt: Scalars['String']['output'];
  deletedAt?: Maybe<Scalars['String']['output']>;
};

export type LinkPublishableApiKeySalesChannel = {
  __typename?: 'LinkPublishableApiKeySalesChannel';
  publishable_key_id: Scalars['String']['output'];
  sales_channel_id: Scalars['String']['output'];
  sales_channel?: Maybe<SalesChannel>;
  createdAt: Scalars['String']['output'];
  updatedAt: Scalars['String']['output'];
  deletedAt?: Maybe<Scalars['String']['output']>;
};

export type LinkRegionPaymentProvider = {
  __typename?: 'LinkRegionPaymentProvider';
  region_id: Scalars['String']['output'];
  payment_provider_id: Scalars['String']['output'];
  region?: Maybe<Region>;
  payment_provider?: Maybe<Payment>;
  createdAt: Scalars['String']['output'];
  updatedAt: Scalars['String']['output'];
  deletedAt?: Maybe<Scalars['String']['output']>;
};

export type LinkSalesChannelStockLocation = {
  __typename?: 'LinkSalesChannelStockLocation';
  sales_channel_id: Scalars['String']['output'];
  stock_location_id: Scalars['String']['output'];
  sales_channel?: Maybe<SalesChannel>;
  location?: Maybe<StockLocation>;
  createdAt: Scalars['String']['output'];
  updatedAt: Scalars['String']['output'];
  deletedAt?: Maybe<Scalars['String']['output']>;
};
// ...
export type LinkProductProductBookModuleBook = {
  __typename?: 'LinkProductProductBookModuleBook';
  product_id: Scalars['String']['output'];
  book_id: Scalars['String']['output'];
  product?: Maybe<Product>;
  createdAt: Scalars['String']['output'];
  updatedAt: Scalars['String']['output'];
  deletedAt?: Maybe<Scalars['String']['output']>;
};